src/requirements.txt
asgiref==3.6.0
attrs==22.2.0
awsebcli==3.20.3
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==39.0.0
Django==4.1.5
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-mptt==0.14.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==3.4
jmespath==0.10.0
jsonschema==3.2.0
paramiko==3.0.0
pathspec==0.9.0
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
PyYAML==5.4.1
qrcode==7.3.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
texttable==1.6.7
urllib3==1.26.14
wcwidth==0.1.9
websocket-client==0.59.0
